Edit TradePaymentInfo to SpecifiedTradePayment and SpecifiedTradePaymentTest
- Add SpecifiedTradePayment
- Add SpecifiedTradePaymentTest

**Change normal form to lombok form model
